Code reorgs - Moved functionally identical code blocks into functions - ProgressBar and LoggerFactory now correctly being disposed - Column headers now work for DiskToBlob
